requirements.txt
absl-py
astunparse
attrs
certifi
charset-normalizer
colorlog
colour
contourpy
cycler
dtreeviz
exceptiongroup
filelock
FLAML
flatbuffers
fonttools
fsspec
gast
google-pasta
graphviz
grpcio
h5py
idna
imbalanced-learn
imblearn
iniconfig
Jinja2
joblib
keras
kiwisolver
libclang
lightgbm
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
ml_dtypes
mpmath
namex
networkx
numpy
nvidia-nccl-cu12
opt_einsum
optree
packaging
pandas
patsy
pillow
pip
pluggy
polars
protobuf
pyarrow
Pygments
pypandoc
pyparsing
pytest
python-dateutil
pytz
PyYAML
requests
rich
river
schedule
scikit-learn
scipy
seaborn
setuptools
six
sklearn-compat
statsmodels
sympy
tensorboard
tensorboard-data-server
tensorflow
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl
tomli
torch
torchaudio
torchvision
tqdm
typing_extensions
tzdata
urllib3
Werkzeug
wheel
wrapt
xgboost